Auditing: Add device.signin event.
This event includes the IP address of the device as seen by the AeroFS
appliance. This will make Bloomberg happy. They will use this event in
combination with splunk's search features to identify when files are
transferred across network boundaries.

Originally I was going to publish this event via Verkehr, but because
verkehr doesn't do any sort of CNAME verification, I determined that was
the wrong place for this to go.

One con of using this event to identify a device's latest IP address
(which is what BB will do) is it relies on the behavior of the ACL
syncronizer, i.e. the behavior that we'll do an ACL sync when we
reconnect and this requires a device.signin. It might be more
appropriate to add a `device.online` event that is not strictly bound to
the ACL syncronizer, and run an OnlineSyncronizer in the daemon. For
now, we simply add this device.signin event, because it accomplishs what
BB wants and besides, this is just a temporary solution anyway, until
the snooping relay is finished. Plus it's only 7 lines. Booyah.

Change-Id: Ifc463d2080beab77c09ff46084381bf646d006e5