Device authorization: add forgotten `device` tag to `device.signin` event.
This is one of those `duh` moments.

I was creating the splunk queries for Bloomberg and I noticed the device
tag was missing.

As far as I can tell there is currently no nice way to enforce audit
behavior in SP via testing. We should consider adding this in the
future.

Change-Id: Ic4710a9244984dd7249055abb9de4caf11ea4714