Windows OpenJDK8 build
Numerous things moved around in the JRE.  I added sunec.dll and the
corresponding jar so that the Windows builds can also use elliptic-curve
cryptography (we already ship this library on OSX).

Most of the file churn is from zoneinfo files moving into a single
timezone database file.  There's a fair bit of additional classes added,
see https://github.com/aerofs/openjdk-trim/pull/1 for the details on
what changed in openjdk-trim to support this commit.

One last thing: since we now use the "server" jvm.dll, we needed to
update the path the launcher searches for the library, so there's a very
small set of launcher changes that need to come along with this.

Change-Id: I13c46f0c6385685eec8038d448d373296b52e965