HttpSessionRemoteAddress: Use X-Real-IP instead of X-Forwarded-For
X-Forwarded-For is a comma-separated list of IP addresses at each step
in a proxy chain. X-Real-IP is set by nginx to be the remote address
that makes the request.

Example proxied call:

1.2.3.4 -> 100.0.0.100 -> 123.123.123.123 -> nginx -> SP

nginx will set these headers:

X-Forwarded-For: 1.2.3.4, 100.0.0.100, 123.123.123.123
X-Real-IP: 123.123.123.123

N.B. the "original" IP in the X-Forwarded-For chain can be easily
forged if, say, 1.2.3.4 set the X-Forwarded-For header in his request to
100.0.0.100

Change-Id: Ied51976eca65911e7e8c6218486aadada0132595