Converted shelob to use bootstrap divs instead of a table
Look and feel is almost identical to the table version--hover color
is a little lighter, last modified column is a little wider for now,
and whole thing is a smidge wider.

This will make adding sub-rows for link-based sharing UI a heck of
a lot easier (you can't surround <tr>s in extra markup, for
instance).

Change-Id: I7bb52884d37adfb599a329877135cafbbd0a2b4f