Resolve a potential ChannelPreallocator deadlock
It turns out that of course registering a listener can call the onComplete
action immediately; and if that guy tries to reach back up and do anything,
it may deadlock.

This merely pulls the allocation of new channels out of the channel directory
monitor, where it should have been all along.

Change-Id: I6c3a208792c28ce1a8b77bcc508e7cd666bf62cb