ENG-1605 avoid heartbeat during cname handshake
CName verification doesn't time out by itself, leaving
upstream handlers to deal with such situations. When
the idle timer fires, close the connection if the CName
handshake hasn't completed.

This does not significantly change the behavior of the
tunnel but makes logs more pleasant to read and easier
to understand by replacing a nasty stack trace with a
single self-explanatory line.

Change-Id: I22b20ab32326e0a4cb66d732f34aa86695a9c8df