Presence changes only with valid channels.
This checkin changes the meaning of a device being present or absent.

Although we will observe and react to both multicast and unicast, only
Unicast will cause an actual device-presence edge transition. This means
the device is marked online when:

 - TCP: Pong received
 - XMPP-using transports: handshake completed.

The device is marked offline when the last connection for a given device
is closed for any reason.

The biggest changes are to the XMPP Presence system and to the ChannelMonitor,
who is responsible for starting and restarting unicast connections based on
the multicast state of a given device.

A future change will be to remove Pulsing; it is no longer needed but it's a
bit of a pain to remove in one go.

AMENDED: fix for xmpp diagnostics
Change-Id: I3273e36396155a3c676cdd0e42154109f5e7143d