Link-based sharing exists, with password protection and expiration dates.
*** On the link creation side: ***

User can create and delete links, and add or remove passwords and
expiration dates to links. A link icon appears in the row for
objects with at least one link.

Password and expiration management are in modals. Custom expiration
dates can be chosen from a datepicker. Expiration is on midnight of
that day--currently no way to choose a custom time. Note that the
datepicker dependency has been modified slightly to fix a bug due
(probably) to differences between BS2 and BS3.

Frontend data model supports having multiple links, though current
UI doesn't make it possible to make more than one link for a given
folder or file (has happened in testing, however).

The success messages are annoying since they move everything
on the page. That's an existing fault, though, that should be fixed
separately.

*** On the link consumption side: ***

When the user goes to a linksharing link, they see a modified
version of Shelob with the file or folder contents being shared.
They can navigate up and down the folder tree, see file names and
sizes, and download any file listed.

If a link is password protected, the user will be prompted to enter
the password before they can see any files.

If a link has expired, the user gets an error notice that the link
either has expired or never existed.

*** Additional changes ***

This commit has shelob display file size, both on My Files
and on linksharing pages.

Individual file rows are now outsourced to a directive, using separate
templates depending on whether it's My Files or a linkshare page.

Change-Id: I4dd30955751a155483ae92e4fb7264de0ecfa264