Expulsion: handle child stores before walk
Preliminary step towards incremental expulsion.

The most important requirement when doing incremental
operations on the object tree is to make sure the first
step achieves the same externally observable effects
as the entire operation, leaving only internal cleanup
to be delt with in subsequent steps.

In the case of expulsion, store bookkeeping must be pulled
out of the tree walk.

Change-Id: Iff01f55ab41a56b0bb290a04cff7768ae389a5a5