Don't filter presence updates in ARP
ARP should send every presence update to the multicast listener.
This will defang the case where an ARP update arrives in the midst
of tearing down the multicast system; previously this meant we would
never tell Presence that a device showed up, thinking that he already
knew.

As a bonus, this removes a bunch of synchronization points from the
ARP class. It's already using a concurrent map with known characteristics,
very rarely will we need to lock it at all.

Resolves ENG-1742.

Change-Id: I6f0021ac779c4af6d96d21f32f60e7faa26cbc14