Normalize shellext-provided path on OSX
One does not simply accept an NFD path from the filesystem.

They are represented as NFC in the db so trying to lookup
with the NFD path yields ExNotFound, leading to much frustration
when trying to share, resolve conflicts or view history for
a path that includes accented characters.

Fixes SUPPORT-1658

Change-Id: I42946de0f24bc00d8e49555acb2a595e8b886323