Remove DevicePresenceListener class.
The DevicePresenceListener would stop all channels for a given DID.
But it would only be signalled when the device went offline...
which would happen when all the channels for a given DID were offline.

So that's a bit circular. And it introduced a potential race condition that
would impact syncing.

Change-Id: Idde09dbe46207d7ec18ad4e266d97d12c55d65c8