Deactivate user scripts: allow support agents to deactivate users.
Lacking a UI that allows users to delete themselves, this is the next
best thing.

Deactivation isn't quite deletion, which is what we tell users we're
doing. Legally we have to provide a way to remove every trace of a user
from our servers. We currently can't do this. Deactivation is a good
first step.

We generally don't run Java jars locally so the bash script is a little
dirty. Potentially we can move this to a support server later or even
attach it to lookup. For now this is fine.

Change-Id: Iea776a69577f35225591b94a7c0ad99e9a4bbe69