TCP store-online notification requires device up
Don't notify core of a store being online just because you heard about it
in Multicast. Ask the canonical Presence class if the device that reported
this 'fact' is actually online. If that device doesn't come online? No stores for you.

This resolves ENG-1871, which manifests as a bad nosync.

Change-Id: Iaa33f65c7912beed9bdf51ef6d828eada5a38a4c