aerofs-db-backup: better sparse file handling
Two things:

rsync does not deal with sparse files very quickly.  Rather than copy
only their contents, it will scan the entire file looking for sequential
blocks of zeros, then make the output file sparse.  While this minimizes
disk usage, it also takes a very long time to make copies, and is slow.
Prefer using cp instead, which will preserve sparseness and only copy
the actual data segments.

Additionally, while bsdtar has great support for producing sparse
archives, this support did not land until version 3.0.4, and the version
in the Ubuntu 12.04 repositories is 3.0.3.  So by using bsdtar at this
juncture, we are in fact forcing all of the sparse files to be expanded.
This is bad and will likely fail.

Change-Id: Ibe59aa1a1f94e4472680c01c0ad9bb384dc10ec8