codesigning: OSX 10.9.5+ is mean
In 10.9.5, Apple obsoleted --resource-rules in favor of requiring all
assets in an application bundle to be signed.  If you add an unsigned
file like site-config.properties, it will break the codesigning seal.
The new behavior is that if you add a resource rule to explicitly
whitelist that file's variability, the linker/loader will refuse to load
your program on the grounds that its resource rules are obsolete.

This has the unfortunate property of breaking our site-config.properties
magic for private cloud repackaging.  PUBLIC needs no repackaging
though, and can use the default resource rules.

We now pick which resource rules to use with codesign based on whether
we're building for PUBLIC or PRIVATE.

This way, we fix things for HC, and avoid regressing for < OS 10.9.5
in PC.

Change-Id: I911edad00cabeb355c7b15e51f5f7e38d7a3cfd9