Handle missing external roots better
If the daemon detects a missing root on startup, move it to the unlinked
roots database - and do so as safely as we know how.

This is part one of a two-parter. The second part includes the UI that
will wrap around this, including all the SanityPoller integration. The
user will have the option to block-and-retry (if they can put the folder
back where it belongs) or proceed (and treat the folder as unlinked).

Tested by deleting and relinking and deleting and relinking an external
folder.

Also some light code cleanup where it was convenient to do so. Some
low-hanging lambda fruit is still out there.

Change-Id: I8c56590631aed3bbe92e700e5d3dff2045fa4884