Handle missing external folders nicely, UI phase
This adds the UI for the singleuser client that pops up when an external
folder is found to be missing. The options we give a user for this are
to leave the folder, or try again in a few seconds.

AMENDED: Tested with TS.
AMENDED: Updated text.
AMENDED: Added CLI code.

Change-Id: If1d3c058f84ada99ebbcfbf96398554b737c84dc