scheduleAtFixedRate -> scheduleWithFixedDelay
scheduleAtFixedRate appears to queue up a bunch of work in the presence
of suspend/resume.  Eep.

This appears to also happen in CharlieClient, where we batch up a ton of
"I'm still here!" calls, and in HealthCheckService, where we batch up a
bunch of deadlock checks, which are high-CPU-usage.

Given the use cases (last seen time, liveness check), it makes more
sense to schedule with fixed delay, which will wait a minute after each
checkin before attempting to check in again, rather than backlogging
multiple state checks.

Change-Id: Ibacbb46ce149b4a53a7e2273380762aa4bb7d4b1
Fixes: ENG-2170