Admin can set organization-wide TFA policy
On the org settings page, an admin can choose whether
to make TFA mandatory, allowed, or disallowed.

Also: various copy and styling changes to TFA pages,
including different template stylings and slight copy
changes in case TFA is mandatory.

Includes double-lock phone cartoon.

Note: right now once a user has logged in, if TFA is mandatory,
they cannot log out again because the button to do so is gone
until they set up TFA. This is...hopefully not a problem for
users in the usual flow, but annoying for testing.

Change-Id: Id25505d5de41bf04af9eb5df5004821f096be7a8