web: rework permissions for 2FA
Add two new privilege groups: two_factor_setup and two_factor_login.

Your request carries the former if SP responds to a request with
ExSecondFactorSetupRequired, or you are a logged in user.  Your session
carries the latter if SP responds to a request with
ExSecondFactorRequired, or you are a logged in user.

This way, we can use the Pyramid privilege system as it was intended.

Change-Id: I750ebdd34f78b5b76b28ae797852915de067fee3