Fix race between RAAU and Linker
As seen on CI.

core.linker.should_not_create_dupliacte_when_move_over_same_content

Download completion before linker picked up filesystem change would
cause the file to be incorrectly marked as cNRO and weirdness would
ensue.

The fix is two-pronged:
  - make RepresentabilityHelper more robust to race conditions
  - fix corner case in LinuxNotifier that could result in missed
    notifications under a freshly re-admitted subtree

Note that the failure seen on CI comes with extra weirdness that
suggests either inotify or the filesystem lost it so the specter
of this failure still lurks.

Change-Id: Ic7cf1cb7fda97298b65553b8df4ead80391e73c7