sp: remove 2FA from sessions when user disables it
Whenever a user's 2FA enforcement is disabled, we revoke the "provided
second factor" bit from all their sessions.  This keeps a session which
had previously been 2FA-proven with a no-longer-valid 2FA secret from
being able to maintain 2FA privileges.

Note that this does NOT revoke 2FA from sessions when 2FA is reset via
Sparta.  Doing so is not currently possible without extracting session
management to a piece that both of the two can talk to, or without
exporting some other interface from SP for session management, neither
of which are currently resourced.

Change-Id: I83ed0f6ea4473379a4dfb0c36543e05213669ed4