linksharing: don't include analytics scripts
Knowing the URI of a linksharing link is sufficient permission to
retrieve the shared content.  As a result, we have to be careful about
not leaking the URI to systems that aren't supposed to have it.

All of our analytics will read the URI and report it to the remote.
This includes:

1) Google Analytics
2) Optimizely
3) Mixpanel
4) Pardot

These file links are then treated as "public" information, legally.
We should stop doing that.

Additional note: we will leak the URI via the Referer: (sic) header to
any host that serves page assets to the linksharing page.  I believe
that with this commit, the leak is limited to cloudfront.

Fixes: ENG-2217
Change-Id: I9871fdcebb8c86aad646acbb9d4bdd3ed8e04036