Group sharing step 10.
Extract shared folder methods into a model because we'll be growing it.

Since we now serialize loading SFM on a single thread, this indirectly
mitigated an issue where the wrong shared folder will be selected after
sharing a folder (caused by a race condition). However, this does not
directly fix the issue and we should examine that issue at a later point.

Change-Id: If3fedba184e8c7426e3fba87d752bf870a88783e