Behold the power of me
All path params that Sparta interprets as a User now
support a "me" shorthand that resolves to the calling
user (as derived from the OAuth token).

Similarly, all path params that are interpreted as a
SharedFolder now support the "root" shorthand.

No changes to existing or future routes required!

The OAuth validation is moved slightly earlier in
request cycle in a container request filter. The
token is then attached to the request context.

The old OAuthProvider just retrieves the token from
the request context.

FactoryReaderProvider is extended to allow factory
methods to accept an optional HttpContext second
parameter. This relies on Jersey's @Context magic.

User and ShareFolder factory methods exploit this
property to transparently handle the special tokens
"me" and "root".

Change-Id: I86cf08d65a6623506ae289ff149bc0e1330fdf05