SV: kill SV references. Clean up support scripts.
The retrace code was in bad shape. It's still not great but at least
it's working now. More cleanup to be done there later, but for now this
is a good improvement.

My best guess is that this was broken when we changed our logging
format. I didn't dive too deep into RCA.

Change-Id: Ia3e7d3f81b69dfd70fb4c1d6c61cbb623e6ff49f