Fixed assorted sharing-related bugs.
on the web client:
- under "My AeroFS" -> "Manage Shared Folders", the UI will allow
the session user to manage a folder when the session user has no
permissions to manage as long as the organization admin does. This
is caused by web back-end sending erroneous data to the front-end.

- if the session user only has permissions to view shared folder
members of a particular folder, the UI will show that all members
of the shared folder having only viewer permissions regardless of
what the actual permissions were.

- the UI will allow the session user to change the session user's
permissions in a shared folder if the session user has permissions
to manage the shared folder.

- when the organization administrator views an organization user's
shared folders. If the user's e-mail contains url-non-safe
characters, then the request will fail and result in a 500.

- when the organization administrator views an organization user's
shared folders. If the user's name contains url-non-safe characters,
then the user's name will not be displayed properly in the
subsequent page.

Change-Id: Ic362ae99276a32382de42d78ed05a49d15790671