No more overwriting CA certs.
Once upon a time in the PC land, the desktop client will read the
CA cert from the config server and override the one provisioned
by the cacert file. While all Java codes use the config property,
overwriting the cacert file was done so libjingle can pick up the
new cacert as well.

We've since discontinued the plan to eventually support jingle in
the PC land, so overwriting the cacert file is no longer necessary.

In addition, overwriting the cacert requires the local user to
have permissions to edit the application binary which is undesirable
in the PC land, hence this removal.

Change-Id: I743fbf877b701ce2bd261fc9798d55cb952574e7