Sparta: Accept non-user-bound tokens for read-only routes
Accepting non-user-bound tokens for routes that alter the DB would
require more thought be given to auditing.

NB: for practical reasons, this commit drops etag support from a
handful of routes. Proper etag support could coneivably be brought
back at a later point but it is not considered important at this
timme.

Change-Id: I227984b283828c8a8a0310f8fdfdb454cdf9d425