Incremental prefix chunking
Or more accurately, online-(split and hash) followed
by background storage to the actual backend.

Block storage prefixes are split into 4Mb chunks as
data is received. Incremental hashing is done for
both the entire file and each individual chunk.

Once the complete prefix is received, the chunks are
moved to an "upload area", from which they can be
retrieved to service read requests until such time
as they are successfully stored in the actual backend,
which happens asynchronously, in a way that is resilient
to daemon crashes.

Change-Id: Ie6d743c6d352244c11f30f83c7e1143f9947fb02