Rework config client for Java servers
The Java servers now use an authenticating config client to request
properties from the config server.  This config client has been
refactored for readability and simplicity.

As a result, Bootstrap can no longer use static globals, since bootstrap
is responsible for generating the secret that it will use to
authenticate to the config server.

To allow bootstrap tasks to defer their dependency on configuration
until they run, a new type ConfigReloadingBootstrapTask is created,
which will fetch and populate the global configuration before running
the deriving task itself.

As a bonus, we no longer use the verkehr hostname to provision the
auditor's certificate.

Change-Id: Id212b8c9a03b92c209b7d75841f5bd8feab86472