Further rework on license sorting
The notion that a paid license is returned in front of an unpaid one
causes new problems; a paying <30-seat customer never sees their
new free 30 seat license.

This adds a model function that returns only an active license, which is
shown on the user's dashboard page *as well as being the target of the
license download*. In order to fix some weird edge cases in
upgrade/downgrade, the legacy newest_license is updated to sort by
modification date and ignore the paid/unpaid thing.

(Otherwise you might be unable to order N seats if you ever had >N seats
paid for. In general, we should count on "most recently touched" as the
best indicator that a given row is "the most relevant license".

Change-Id: I496005c68274cc316c630ff429a9f64b66bc2493