Bifrost: make all routes safe to expose publicly
We achieve this by requiring Aero-Service-Shared-Secret authentication
for any route that exposes information or functionality that wouldn't be
safe to expose to the world.

Some of the interactions between web and bifrost would, in an ideal
world, use Aero-Delegated-User authentication.  However, Bifrost
currently lacks the ability to answer whether a delegated user is an
org admin, and who they are allowed to manage, and this is not exposed
by Sparta in a straightforward manner at present; as a result, we
instead rely on web to perform access control checks, and use the shared
secret authenticator instead.

The /tokenlist?owner= route is removed in favor of
/users/{userid}/tokenlist because it better fits a REST noun hierarchy.

Change-Id: I7abefceb5c560cb505f536bd849836496409b719