Manual ACL inheritance for new files
Creating a dummy file to use as target to ReplaceFile is
brittle as we run the risk of some other process detecting
its creation and obtaining a lock on it such that it
becomes impossible for ReplaceFile to proceed.

In addition, in such a case, the dummy file would not be
cleaned up properly, which could easily result in a race
wherein the scanner detects it and the next download
attempt results in the creation of a conflict branch.

As seen in CI.

Also replace a few uses of SytemUtil.fatal with throwing
Error to allow finally blocks to be processed.

Change-Id: I4873e28dd122c3a1023e470aa801fb372b9ed0da