acl checks on admins use sf's existing members
organization admins are allowed to change the ACLs of any shared
folder in which a member of their organization is a pending, left or
joined owner.

also check that the user exists when checking this condition so we
don't hit exnotfound errors. though technically not necessary in PC,
imo the impact there isn't big enough to make the logic more complex

fixes SUPPORT-2081

Change-Id: I5ece9a33136a3e205b4edc53cac04b0edc70ac56