shelob: show link options only when user has privileges
Previously, the user would see the option to create a link for every object in
the view, and attempting to create a link for an object for which the user does
not have MANAGE privileges would cause an error.

Now the option is shown only on the objects for which link creation would succeed.

Change-Id: I3605e5261e8507bd76c2df1576b70d9a44119ce3