[greasefire] charlie
Although Charlie is not itself massively fat, it is the only user
of persistent vk topics and therefore stands of the way of vk being
replaced by lipwig.

Client API is unchanged (although support for the old Authorization
format is dropped).

A new (shared-secret-auth'd) route is introduced to replace vk topic
reads. The JSON payload format is preserved but the layers of vk
encoding disappear.

Data is stored in a boltdb database using a fairly compact binary
encoding (16 bytes key, 12 or 24 bytes values for ipv4 and ipv6
checkin addresses respectively).

Update web and sparta to query charlie instead of vk.

Change-Id: Ic27e3074305e011477c52fb097933de230e8dd91