Fix IncomingStream#read()
Damn java and its lack of unsigned types.

Byte values were implicitly converted to int w/ sign extension,
turning 0xff into -1, the special value used to denote stream end.

As a result, any length-delimited protobuf message whose serialized
size was a 1 short of a multiple of 128 would result in decoding
returning a null.

This was observed with suprising frequency in the wild, where any
such PBGetVersionsResponseBlock in a GetVersionsResponse would cause
an NPE, leading to a no-sync for the affected store.

Change-Id: I1c64405c8b9265be9a3f82f538876692846be4d3