Maintenance: render maintenance mode page from maintenance-web, not bunker.
Bunker is not accessible to the general public, so in an enterprise
deployment what the end user used to see during maintenance mode is an "unable
to connect to share.acme.com:8484", which isn't great.

This commit introduces maintenance-web, which is a slimmed version of
web that includes only the maintenance mode page. Sadly, there is a
bunch of boilerplate code here. IMO this the lessor two evils, the other
evil being running web in a different mode. Code sharing is done where
possible, for example with static assets (similar to how bunker shares static
assets with web).

Resolves ENG-2616.

Change-Id: Ia81e6cb307b096581ca050d07ea56ecdcca7eace