Fix GUI not showing InitErrors.
https://gerrit.arrowfs.org/#/c/4040/ broke a feature where the GUI program
shows user error messages should the configuration values failed to load.

The cause is that cacert is now being read at initialization instead of
on-demand and the configuration is not guaranteed to have loaded
successfully at initialization time.

The fix is to ensure either configuration is fully loaded and ready or we
short-circuit straight to display errors.

Change-Id: I7dfc425cddb5a859d1e028fc2dac407e59ad0e35