Fix NPE
Local modification can abort a download causing the
prefix to remain even though no prefix version was
stored, if the content is small enough to fit in a
single unicast packet. In that case the prefix should
not be used when making the next content request.

Also adapt content conflict tests to work with Phoenix.

Change-Id: I3ab74adc8571fcd268368380f99df124c444e3e5