Misc fixes
Race condition in TokenVerificationClient could close the
connection before the request was given a reasonable amount
of time to complete, resulting in an incorrect 401 response.

Regression in content conflict test harness for non-phoenix
setups (typo).

Fix incorrect use of network partition in a shared folder
deletion test which caused it to timeout for phoenix clients.

Change-Id: I9e8b822d5659306b189a9f8448004751adba3b11