packaging overhaul
Completely overhaul how packaging/repackaging works. This patchset
introduces the "updater" service, which acts as a replacement for
static file distribution; this updater becomes the aerofs binary
and is responsible for pulling down the sources.

We also modify the build process for the repackaging image, removing
a large amount of hard-to-test logic at the expense of half a dozen MBs
of extra space in the container. This helps allows us to more easily
test the real changes to the nginx container, eg. that it serves the
static sources.

This patchset roughly halves the time needed to package clients (as
well as only rebuilding them when necessary, thus speeding up rebuilds
by up to a minute) and, more importantly, lowers the time needed to
repackage from roughly 2.5 mins to just over ten seconds. The
repackaging container has also been halved in size (down ~600MB),
while the nginx container has increased by ~200MB.

Finally, these changes result in much faster client updates: since
the binary itself a) will rarely change, and b) is much smaller,
updates, when necessary, require a much smaller download. Most old
static files are simply reused. This also paves the way for more
seamless updates; if the updater itself does not change, static files
could be updated without specifically replacing the binary. This
capability is not included in this patchset, but may be worth doing
in the future.

Change-Id: Ia0832c0ac5169b8735b7d1acfedd54b541d35592