ENG-3027: no-sync after unlink/reinstall
Problem
-------

When a user shares a folder, their root store has
transforms for the creation of the whole subtree
then a SHARE op. Any further changes happen in the
shared folder.

In contrast, when one joins a folder, the root store
only sees the creation of an STORE/ANCHOR (in polaris
and daemon respectively).

When one unlinks and reinstall, shared folders are
restored if valid tag files are found during the first
scan. This results in the immediate creation of an
ANCHOR and the corresponding store.

When such a restored client would receive transforms
for its root store leading to the creation of the shared
folder it would fail to perform the SHARE op as the
anchor was already present.

Solution
--------

Detect folder/anchor conflict when receinving INSERT_CHILD
transforms and mark the object as buffered.

When applying buffered changes, drop buffered folder with
a matching anchor and all its buffered children.

SHARE transform is a no-op if the anchor already exists

Change-Id: I6b8ffa8e802f06bd3c9f8974d37faed6a8843f77