Api documentation using swagger-ui
This adds swagger specs for some existing services.

Also, a script that generates a simple index.html with links
to each swagger spec as opened in swagger-ui. This script will be
integrated in the docs-server container in the success-utils repo.

Also, a fix and some improvements for the markdown_watch script.

Change-Id: I7843487b4c94b8ae6c23e9528065290d2cb8b872