Automate generating an AeroIM combined appliance
Use the build-images and modify-appliance scripts to our advantage to
import the needed images, and enable AeroIM in the CI and local-prod
cases.

To release AeroIM to the wild, simply move messaging from sites-disabled
to sites (under docker/nginx) and replace the checked-in crane.yml.

This version imports the needed IM containers from registry.aerofs.com.
A better (git-submodule workalike?) way could be devised when these
codelines start to approach integration.

Change-Id: Icea03000ca0ab1a73575e92d5af397256b251485