Python unit tests: fix requirements.txt. ENG-3168.
As it's written requirements.txt soft links to requirements-exact.txt.
This is not the best way to do things; requirements.txt + pip freeze
should be used to generate requirements-exact.

That's orthogonal to this fix though. Now web, maintenance-web, and
bunker look the same and the tests will pass. We can clean up the above
in a future commnt.

Change-Id: I80ffab6e93078ac7c9f4a2a25c551e04a9417737