Update trifrost in accordance with prophecy
The identity-management document describes the overall plan.

Trifrost is an identity-verification service. It checks that a
supplicant actually owns the email they claim. Based on that, it makes a
call to Sparta/Bifrost to get an access token.

This commit includes a new route on sparta to generate tokens for
services that delegate user actions.

Trifrost is being gutted; many tests died to bring you this smaller
service.

Change-Id: I6734daa30ea59f7bde108852f5a72be4f2d2f1cc