Lizard: Use Flask-KVSession and ext to prevent cookie replay. ENG-3249.
Flask-Login is cookie based and so there's no way to destory a session
for a given user. With Flask-KVSession, we back our session data in
redis and only pass a unique ID to the browser cookie. Therefore, we are
able to destroy sessions server side and prevent replay attacks when
users reset their passoword.

In the Flask-KVSession docs, they assert "[with Flask-KVSession] it is
possible to securely destroy sessions to protect against replay attacks."
Practically speaking, however, they make this _very_ hard to do, as
sessions are not indexed by user in any way under the hood with FKVS.
So, we need to roll our own wrapper to obtain a user => session list
mapping. I have done this in login_helper.py.

The solution here isn't the cleanest, but it works and it's the easiest
solution I can find that plays nicely with Flask.

Change-Id: Ia8fca8aba48d19011125ee4345d150b8750dde61