Restore Activity Log post-phoenix
This is a stopgap. Activity Log should eventually be overhauled to reflect
the way Phoenix works. Crucially, it would probably make more sense to
have activity log on the web hooked directly to the transforms feed on
polaris. It would also be good to fix AL to reflect high-level operations
such as sharing/migration instead of showing a confusing jumble of delete
and create operations. Lastly, if we are to keep a local activity log
on the client we should consider highlighting the difference between
accepted remote changes and pending local changes.

Change-Id: Id633743bfbf4a64685463c0551422c2ab364c712