Phoenix Conversion: cleanup conflicts
Change of conflict-related invariants require pre-phoenix conflicts
to be cleaned up to avoid issues post-phoenix.

This has the unfortunate effect that conflict branches will have to
be re-downloaded post-conversion. In some rare cases it might also
force users to dive into sync history to retrieve conflict branches.

To avoid out-of-date clients from incorrectly treating newer content
as conflicting post-conversion, a dummy central version must be added
for every object with local content. Ideally this would be done with
some hints from polaris to preserve true conflicts but this is not
trivial and would also not work for Prominence since they've already
been through conversion. The simplest solution is thus to always add
it, which may result in true conflicts being silently moved to sync
history, a suboptimal but safe behavior.

Lastly, to account for the fact that Prominence already converted and
may have generated fresh post-conversion conflicts, the DPUT used to
cleanup pre-conversion ones needs to take some care to distinguish
between the two.

Change-Id: I0da063b6e1a3814d800e790dff7900b84d14877d