Conflict UX improvement
Automatically cleanup conflict branch when linker detects that
MASTER has matching hash.

Users would intuitively use "Save As" to overwrite the local
version with a remote version and expect it to resolve the
conflict but until now it resulted in two branches with the
same content hash. Now the conflict branch is cleaned up, as
expected by the user.

NB: the change also applies pre-phoenix but not quite as
optimally as one might hope. If there are multiple branches
for a given file, only the matching one will be cleaned up.
The user still need to delete every other one to completely
resolve the conflict.

Change-Id: Id0e41af0db75bcbf5b33f3955364c155fd8feaed