sloth: discard closed websocket connections
This commit fixes a bug which caused websocket connections to accumulate
in the broadcaster, eventually slowing things down and causing sloth to
timeout.

The Gorilla websocket docs specify that reads are necessary on each
websocket connection to handle ping and close messages.

https://godoc.org/github.com/gorilla/websocket#hdr-Read_is_Required

This commit listens to each ws connection, discarding all messages and
unsubscribing when the connection is closed.

Change-Id: Ice52cdf5791b93969dcba158716b0156664a7a96