sloth: introduce data messages
This commit introduces the concept of "data" messages in a conversation,
which represent events that should be viewable to a user in the message
thread, but are not traditional messages written by a user to a user
or group.

An example is group membership history changes. Previously, events like
"JGray left the conversation" were stored in a separate db table,
fetched by the client through a separate API call, and interleaved with
the "traditional" messages client-side. This complicated the client code
significantly.

Group membership changes are now stored in the messages table with a new
"is_data" column set to true. The body is a json blob with the "type"
key set to "MEMBER_ADDED" or "MEMBER_REMOVED".

Change-Id: Ibcd68807591924738355fb040dd014c9b3888ac9