ENG-3715: Lizard: send email when attempting to create duplicate account.
Make sure to handle the promo code case. This case is particularly
important because with the upcoming marketing campaign, we'll be
directing 80k people to a page with a form for new users and a link for
existing users. Undoubtedly, existing users will try to use the form.
Let's make sure this thing is foolproof.

Change-Id: Ia8c7d0a77a1a1537d7d551b655de238ca4fa133a