ENG-3791 Dummy conflicts
When remote content is known but unavailable, either temporarily
or permanently, as can happen when devices go permanently offline
or when an appliance is restored from an outdated backup, local
changes will permanently be rejected by Polaris.

Because a single failure currently blocks any subsequent content
submission despite their independence (something which should be
addressed separately), this causes not just a no-sync for the
affected file but for all changes in the same shared folder.

To solve this, the invariant around conflict branches are changed
to allow the creation of "dummy" conflict branches, which do not
correspond to any centrally accepted version. This allows the issue
to be reported to the user through the existing conflict resolution
UI, at which point action can be taken to resolve it by picking the
local version over the unavailable remote one.

Two syncdet tests are added to exercise this new behavior and the
conflict dialog is adapted to correctly handle dummy conflicts.

Change-Id: I037edd0a33f6a8484dbf8e500d922ce0ccd7a51c