ENG-3843 Fix no-sync/crash caused by share+unlink
When a client is unlinked and reinstalled (or seeded w/ tag files)
anchor object for all shared folders will be created before the
client receives the INSERT_CHILD and SHARE ops for the original
folder.

Previously we would rely on the meta buffer and simply drop changes
in the original folder.

This however fails in specatular fashion when there are multiple
shared folders and changes in some of the original folders are
received after the first SHARE operation. The new changes will
fail to be applied because the target objects were dropped,
resulting either in a persistent ExNotFound or, worse in a
persistent AssertionError.

Dropping these new changes is tempting but it opens the door to
even nastier corner cases.

Instead we now accept all these changes and populate the subtree
rooted at the original folder under the trash, which ensures that
all intermediate changes are correctly applied and in particular
objects can be moved in and out of the subtree before the SHARE
op is eventually applied.

Once the SHARE op is received, the subtree is erased from the
database as would have happened had the anchor not been already
present.

Change-Id: I5523c8f04ffac77c26bbfd38a533235d690666f9